Fix(2). 777929 Token retreival setup. Team's review.
