+++ src/main/java/faang/school/notificationservice/messaging/telegram/NotificationTelegramBot.java
@@ -1,0 +1,43 @@
+package faang.school.notificationservice.messaging.telegram;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+import org.telegram.telegrambots.bots.TelegramLongPollingBot;
+import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
+import org.telegram.telegrambots.meta.api.objects.Update;
+import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
+
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+@Component
+public class NotificationTelegramBot extends TelegramLongPollingBot {
+    private final String token;
+    
+    @Value("${spring.telegram.bot-name}")
+    private String botName;
+    
+    public NotificationTelegramBot(@Value("${spring.telegram.token}") String token) {
+        super(token);
+        this.token = token;
+    }
+
+    @Override
+    public String getBotUsername() {
+        return botName;
+    }
+
+    @Override
+    public void onUpdateReceived(Update update) {
+    }
+
+    public void send(String usersChat, String message) {
+        SendMessage msg = new SendMessage(usersChat, message);
+        try {
+            execute(msg);
+            log.info("Message to user {} was sent.", usersChat);
+        } catch (TelegramApiException e) {
+            e.printStackTrace();
+        }
+    }
+}
